Fix bug

+++ front_arduino/front_arduino.ino
@@ -18,8 +18,8 @@ Servo trusterBR;
 DHT11 dht11(temperature_pin);
 
 // Signal value for truster to move forward or backward
-byte forward_max = 200; 
-byte backward_max = -200;
+int forward_max = 200; 
+int backward_max = -200;
 byte noMove = 1500;
